Don't allow to remove a user that is currently restaurant admin

+++ src/FoodOrderSystem.Domain/Commands/RemoveUser/RemoveUserCommandHandler.cs
@@ -1,18 +1,22 @@
 ﻿using FoodOrderSystem.Domain.Model;
 using FoodOrderSystem.Domain.Model.User;
 using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using FoodOrderSystem.Domain.Model.Restaurant;
 
 namespace FoodOrderSystem.Domain.Commands.RemoveUser
 {
     public class RemoveUserCommandHandler : ICommandHandler<RemoveUserCommand, bool>
     {
         private readonly IUserRepository userRepository;
+        private readonly IRestaurantRepository restaurantRepository;
 
-        public RemoveUserCommandHandler(IUserRepository userRepository)
+        public RemoveUserCommandHandler(IUserRepository userRepository, IRestaurantRepository restaurantRepository)
         {
             this.userRepository = userRepository;
+            this.restaurantRepository = restaurantRepository;
         }
 
         public async Task<Result<bool>> HandleAsync(RemoveUserCommand command, User currentUser, CancellationToken cancellationToken = default)
@@ -28,6 +32,14 @@ namespace FoodOrderSystem.Domain.Commands.RemoveUser
 
             if (command.UserId == currentUser.Id)
                 return FailureResult<bool>.Create(FailureResultCode.CannotRemoveCurrentUser);
+
+            var restaurants = await restaurantRepository.FindByUserIdAsync(command.UserId, cancellationToken);
+            var restaurantList = restaurants.ToList();
+            if (restaurantList.Any())
+            {
+                return FailureResult<bool>.Create(FailureResultCode.UserIsRestaurantAdmin,
+                    string.Join(", ", restaurantList.Select(en => en.Name)));
+            }
 
             await userRepository.RemoveAsync(command.UserId, cancellationToken);
